+++ src/Exceptions/ValidationException.php
@@ -9,7 +9,7 @@ use Lloricode\CheckDigit\Enums\Format;
 
 class ValidationException extends Exception
 {
-    public static function length($value, int $actualLength, Format $format): self
+    public static function length(string $value, int $actualLength, Format $format): self
     {
         return new self("Invalid length of `$actualLength` for format `$format->value`, with value `$value`");
     }