+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to `check-digit` will be documented in this file
 
 ## unreleased
 
-- Allow using pest v2
+- Allow using pest v2.
+- Install phpstan
+- Fix issue using phpstan.
 
 ## 3.0.0 - 2023-10-09